Rename QueryTypes to DataModels
